updated component and ui
